+++ app/src/main/kotlin/com/woowla/ghd/eventbus/Event.kt
@@ -1,7 +1,0 @@
-package com.woowla.ghd.eventbus
-
-enum class Event {
-    SETTINGS_UPDATED,
-    REPO_TO_CHECK_UPDATED,
-    SYNCHRONIZED,
-}