+++ core/src/main/kotlin/com/woowla/ghd/AppFolderFactory.kt
@@ -1,0 +1,23 @@
+package com.woowla.ghd
+
+import java.nio.file.FileSystems
+import java.nio.file.Path
+import java.nio.file.Paths
+import net.harawata.appdirs.AppDirsFactory
+
+class AppFolderFactory(isDebug: Boolean, debugAppFolder: String) {
+    val folder: Path = if (isDebug) {
+        Paths.get(FileSystems.getDefault().getPath("").toAbsolutePath().parent.toString(), debugAppFolder)
+    } else {
+        val appDirs = AppDirsFactory.getInstance()
+        val pathStr = appDirs.getUserDataDir("ghd", null, "woowla")
+        Paths.get(pathStr)
+    }
+
+    suspend fun createFolder() {
+        val file = folder.toFile()
+        if (!file.exists()) {
+            file.mkdirs()
+        }
+    }
+}