+++ domain-api/src/main/kotlin/com/woowla/ghd/domain/entities/RateLimit.kt
@@ -1,0 +1,14 @@
+package com.woowla.ghd.domain.entities
+
+import kotlinx.datetime.Instant
+
+data class RateLimit(
+    val limit: Long?,
+    val remaining: Long?,
+    val used: Long?,
+    val reset: Instant?,
+    val resource: String?,
+) {
+    val percentageUsed: Int? = if (limit == null || used == null) { null } else { ((used * 100) / limit).toInt() }
+    val percentageRemaining: Int? = if (limit == null || remaining == null) { null } else { ((remaining * 100) / limit).toInt() }
+}