+++ domain-api/src/main/kotlin/com/woowla/ghd/domain/entities/SyncResult.kt
@@ -1,0 +1,14 @@
+package com.woowla.ghd.domain.entities
+
+import kotlin.time.Duration
+import kotlinx.datetime.Instant
+
+data class SyncResult(
+    val id: Long = 0,
+    val startAt: Instant,
+    val endAt: Instant?,
+) {
+    enum class Status { SUCCESS, WARNING, ERROR, CRITICAL }
+
+    val duration: Duration? = endAt?.minus(startAt)
+}